Added Basic AI + Minor Glitch Fix
Added a basic AI that just picks a random square.

Fixed a glitch where GameOver could be called multiple times if multiple
end conditions were met. GameOver will now do nothing if you call it
when the game is already in GameState.FINISHED state.

Squares will no longer highlight if they are occupied or if the game is
over.